+++ tests/testthat/teardown.R
@@ -1,0 +1,1 @@
+options(pre_test_options)