+++ tests/testthat/test-parsing-factors.R
@@ -61,7 +61,7 @@ test_that("Factors handle encodings properly (#615)", {
 
   x <- read_csv(f,
     col_types = cols(col_factor(c("A", "\uC4"))),
-    locale = locale(encoding = "latin1"), progress = FALSE)
+    locale = locale(encoding = "latin1"))
 
   expect_s3_class(x$test, "factor")
   expect_equal(x$test, factor(c("A", "\uC4")))