+++ tests/testthat/test-read-csv.R
@@ -1,49 +1,49 @@
 test_that("read_csv col imputation, col_name detection and NA detection works", {
-  test_data <- read_csv(test_path("basic-df.csv"), col_types = list(), col_names = TRUE, progress = FALSE)
+  test_data <- read_csv(test_path("basic-df.csv"), col_types = list(), col_names = TRUE)
   expect_equal(unname(unlist(lapply(test_data, class))),
     c("logical", "numeric", "numeric", "character"))
   expect_equal(names(test_data), c("a", "b", "c", "d"))
   expect_equal(sum(is.na(test_data$d)), 1)
 
-  test_data2 <- read_csv("basic-df.csv", col_types = list(a = "l", b = "d", c = "d", d = "c"), col_names = TRUE, progress = FALSE)
+  test_data2 <- read_csv("basic-df.csv", col_types = list(a = "l", b = "d", c = "d", d = "c"), col_names = TRUE)
   expect_equal(test_data, test_data2)
 })
 
 test_that("read_csv's 'NA' option genuinely changes the NA values", {
-  expect_equal(read_csv(I("a\nz"), na = "z", progress = FALSE)$a, NA)
+  expect_equal(read_csv(I("a\nz"), na = "z")$a, NA)
 })
 
 test_that("read_csv's 'NA' option works with multiple NA values", {
-  expect_equal(read_csv(I("a\nNA\nmiss\n13"), na = c("13", "miss"), progress = FALSE)$a,
+  expect_equal(read_csv(I("a\nNA\nmiss\n13"), na = c("13", "miss"))$a,
                c("NA", NA, NA))
 })
 
 test_that('passing character() to read_csv\'s "NA" option reads "" correctly', {
-  expect_equal(read_csv(I("a\nfoo\n"), na = character(), progress = FALSE)$a, "foo")
+  expect_equal(read_csv(I("a\nfoo\n"), na = character())$a, "foo")
 })
 
 test_that("passing \"\" to read_csv's 'NA' option reads \"\" correctly", {
-  expect_equal(read_csv(I("a,b\nfoo,bar\nfoo,\n"), na = "", progress = FALSE)$b, c("bar", NA))
+  expect_equal(read_csv(I("a,b\nfoo,bar\nfoo,\n"), na = "")$b, c("bar", NA))
 })
 
 test_that("changing read_csv's 'quote' argument works correctly", {
-  test_data <- read_csv("basic-df.csv", col_types = NULL, col_names = TRUE, progress = FALSE)
+  test_data <- read_csv("basic-df.csv", col_types = NULL, col_names = TRUE)
   test_data_singlequote <- read_csv("basic-df-singlequote.csv", quote="'")
   expect_identical(test_data, test_data_singlequote)
 })
 
 test_that("read_csv's 'skip' option allows for skipping'", {
-  test_data <- read_csv("basic-df.csv", skip = 1, progress = FALSE)
+  test_data <- read_csv("basic-df.csv", skip = 1)
   expect_equal(nrow(test_data), 9)
 })
 
 test_that("read_csv's 'skip' option allows for skipping when no header row is present'", {
-  test_data <- read_csv("basic-df.csv", skip = 1, col_names = FALSE, progress = FALSE)
+  test_data <- read_csv("basic-df.csv", skip = 1, col_names = FALSE)
   expect_equal(nrow(test_data), 10)
 })
 
 test_that("read_csv's 'n_max' allows for a maximum number of records and does not corrupt any", {
-  test_data <- read_csv("basic-df.csv", n_max = 7, progress = FALSE)
+  test_data <- read_csv("basic-df.csv", n_max = 7)
   expect_equal(nrow(test_data), 7)
   expect_equal(sum(is.na(test_data)), 0)
 })
@@ -61,20 +61,20 @@ test_that("can read more than 100 columns", {
   x <- as.data.frame(matrix(rbinom(300, 2, .5), nrow = 2))
   y <- format_csv(x)
 
-  expect_equal(ncol(read_csv(I(y), progress = FALSE)), 150)
+  expect_equal(ncol(read_csv(I(y))), 150)
 })
 
 test_that("encoding affects text and headers", {
   skip_on_os("solaris")
 
-  x <- read_csv("enc-iso-8859-1.txt", locale = locale(encoding = "ISO-8859-1"), progress = FALSE)
+  x <- read_csv("enc-iso-8859-1.txt", locale = locale(encoding = "ISO-8859-1"))
   expect_identical(names(x), "fran\u00e7ais")
   expect_identical(x[[1]], "\u00e9l\u00e8ve")
 })
 
 test_that("nuls are dropped with a warning", {
   skip_if_edition_second()
-  expect_warning(x <- read_csv("raw.csv", progress = FALSE))
+  expect_warning(x <- read_csv("raw.csv"))
   expect_equal(n_problems(x), 1)
   expect_equal(x$abc, "ab")
 })
@@ -96,42 +96,42 @@ test_that("warnings based on number of columns (not output columns)", {
   skip_if_edition_second()
 
   # FIXME: the column name when skipping is the wrong name
-  out1 <- read_csv(I("1,2,3\n4,5,6"), "z", "__i", progress = FALSE)
-  out2 <- read_csv(I("1,2,3\n4,5,6"), FALSE, cols_only(X3 = "i"), progress = FALSE)
+  out1 <- read_csv(I("1,2,3\n4,5,6"), "z", "__i")
+  out2 <- read_csv(I("1,2,3\n4,5,6"), FALSE, cols_only(X3 = "i"))
 
   expect_equal(n_problems(out1), 0)
   expect_equal(n_problems(out2), 0)
 })
 
 test_that("missing last field generates warning", {
-  expect_warning(out <- read_csv(I("a,b\n2"), progress = FALSE))
+  expect_warning(out <- read_csv(I("a,b\n2")))
   expect_equal(problems(out)$expected, "2 columns")
 })
 
 test_that("missing lines are skipped without warning", {
   skip_if_edition_second()
    # first
-  expect_silent(out <- read_csv(I("a,b\n\n\n1,2"), progress = FALSE))
+  expect_silent(out <- read_csv(I("a,b\n\n\n1,2")))
 
    # middle
-  expect_silent(out <- read_csv(I("a,b\n1,2\n\n\n2,3\n"), progress = FALSE))
+  expect_silent(out <- read_csv(I("a,b\n1,2\n\n\n2,3\n")))
 
   # last (trailing \n is ignored)
-  expect_silent(out <- read_csv(I("a,b\n1,2\n\n\n"), progress = FALSE))
+  expect_silent(out <- read_csv(I("a,b\n1,2\n\n\n")))
 })
 
 test_that("warning lines are correct after skipping", {
   skip_if_edition_second()
 
-  expect_warning(out1 <- read_csv(I("v1,v2\n1,foo"), col_types = "ii", progress = FALSE, lazy = FALSE))
-  expect_warning(out2 <- read_csv(I("#foo\nv1,v2\n1,foo"), col_types = "ii", comment = "#", progress = FALSE, lazy = FALSE))
+  expect_warning(out1 <- read_csv(I("v1,v2\n1,foo"), col_types = "ii", lazy = FALSE))
+  expect_warning(out2 <- read_csv(I("#foo\nv1,v2\n1,foo"), col_types = "ii", comment = "#", lazy = FALSE))
 
   expect_equal(problems(out1)$row, 1)
 
   expect_equal(problems(out2)$row, 1)
 
-  expect_warning(out3 <- read_csv(I("v1,v2\n\n1,2\n\n3,4"), col_types = "i", progress = FALSE))
-  expect_warning(out4 <- read_csv(I("v1,v2\n#foo\n1,2\n#bar\n3,4"), col_types = "i", comment = "#", progress = FALSE))
+  expect_warning(out3 <- read_csv(I("v1,v2\n\n1,2\n\n3,4"), col_types = "i"))
+  expect_warning(out4 <- read_csv(I("v1,v2\n#foo\n1,2\n#bar\n3,4"), col_types = "i", comment = "#"))
 
   expect_equal(problems(out3)$row, c(1, 2))
 
@@ -140,10 +140,10 @@ test_that("warning lines are correct after skipping", {
 
 test_that("extra columns generates warnings", {
   skip_if_edition_second()
-  expect_warning(out1 <- read_csv(I("a,b\n1,2,3\n"), progress = FALSE, lazy = FALSE))
-  expect_warning(out2 <- read_csv(I("a,b\n1,2,3"), col_types = "ii", progress = FALSE, lazy = FALSE))
-  #expect_warning(out3 <- read_csv(I("1,2,3\n"), c("a", "b"), progress = FALSE, lazy = FALSE))
-  expect_warning(out4 <- read_csv(I("1,2,3\n"), c("a", "b"), "ii", progress = FALSE, lazy = FALSE))
+  expect_warning(out1 <- read_csv(I("a,b\n1,2,3\n"), lazy = FALSE))
+  expect_warning(out2 <- read_csv(I("a,b\n1,2,3"), col_types = "ii", lazy = FALSE))
+  #expect_warning(out3 <- read_csv(I("1,2,3\n"), c("a", "b"), lazy = FALSE))
+  expect_warning(out4 <- read_csv(I("1,2,3\n"), c("a", "b"), "ii", lazy = FALSE))
 
   expect_equal(problems(out1)$expected, "2 columns")
   expect_equal(problems(out2)$expected, "2 columns")
@@ -153,11 +153,11 @@ test_that("extra columns generates warnings", {
 
 test_that("too few or extra col_types generates warnings", {
   skip_if_edition_second()
-  expect_warning(out1 <- read_csv(I("v1,v2\n1,2"), col_types = "ii", progress = FALSE, lazy = FALSE))
+  expect_warning(out1 <- read_csv(I("v1,v2\n1,2"), col_types = "ii", lazy = FALSE))
   expect_equal(problems(out1)$expected, "1 columns")
   expect_equal(  problems(out1)$actual, "2 columns")
 
-  expect_warning(out2 <- read_csv(I("v1,v2\n1,2"), col_types = "iii", progress = FALSE, lazy = FALSE))
+  expect_warning(out2 <- read_csv(I("v1,v2\n1,2"), col_types = "iii", lazy = FALSE))
   expect_equal(ncol(out2), 2)
 })
 
@@ -165,7 +165,7 @@ test_that("too few or extra col_types generates warnings", {
 
 test_that("decimal mark automatically set to ,", {
   expect_message(
-    x <- read_csv2(I("x\n1,23"), progress = FALSE),
+    x <- read_csv2(I("x\n1,23")),
     if (default_locale()$decimal_mark == ".") "decimal .*grouping mark" else NA)
   expect_equal(x[[1]], 1.23)
 })
@@ -173,14 +173,14 @@ test_that("decimal mark automatically set to ,", {
 # Zero rows ---------------------------------------------------------------
 
 test_that("header only df gets character columns", {
-  x <- read_csv(I("a,b\n"), progress = FALSE)
+  x <- read_csv(I("a,b\n"))
   expect_equal(dim(x), c(0, 2))
   expect_equal(class(x$a), "character")
   expect_equal(class(x$b), "character")
 })
 
 test_that("n_max 0 gives zero row data frame", {
-  x <- read_csv(I("a,b\n1,2"), n_max = 0, progress = FALSE)
+  x <- read_csv(I("a,b\n1,2"), n_max = 0)
   expect_equal(dim(x), c(0, 2))
   expect_equal(class(x$a), "character")
   expect_equal(class(x$b), "character")
@@ -189,7 +189,7 @@ test_that("n_max 0 gives zero row data frame", {
 test_that("empty file with col_names and col_types creates correct columns", {
   skip_if_edition_second()
 
-  x <- read_csv(datasource_string("", 0), c("a", "b"), "ii", progress = FALSE)
+  x <- read_csv(datasource_string("", 0), c("a", "b"), "ii")
   expect_equal(dim(x), c(0, 2))
   expect_equal(class(x$a), "integer")
   expect_equal(class(x$b), "integer")
@@ -206,20 +206,20 @@ test_that("empty file returns an empty tibble", {
 
 test_that("comments are ignored regardless of where they appear", {
 
-  out1 <- read_csv(I('x\n1#comment'),comment = "#", progress = FALSE)
-  out2 <- read_csv(I('x\n1#comment\n#comment'), comment = "#", progress = FALSE)
-  out3 <- read_csv(I('x\n"1"#comment'), comment = "#", progress = FALSE)
+  out1 <- read_csv(I('x\n1#comment'),comment = "#")
+  out2 <- read_csv(I('x\n1#comment\n#comment'), comment = "#")
+  out3 <- read_csv(I('x\n"1"#comment'), comment = "#")
 
   expect_equal(out1$x, 1)
   expect_equal(out2$x, 1)
   expect_equal(out3$x, 1)
 
-  out4 <- read_csv(I('x,y\n1,#comment'), comment = "#", col_types = "cc", progress = FALSE)
+  out4 <- read_csv(I('x,y\n1,#comment'), comment = "#", col_types = "cc")
   expect_equal(out4$y, NA_character_)
 
-  expect_warning(out5 <- read_csv(I("x1,x2,x3\nA2,B2,C2\nA3#,B2,C2\nA4,A5,A6"), comment = "#", progress = FALSE, lazy = FALSE))
-  expect_warning(out6 <- read_csv(I("x1,x2,x3\nA2,B2,C2\nA3,#B2,C2\nA4,A5,A6"), comment = "#", progress = FALSE, lazy = FALSE))
-  expect_warning(out7 <- read_csv(I("x1,x2,x3\nA2,B2,C2\nA3,#B2,C2\n#comment\nA4,A5,A6"), comment = "#", progress = FALSE, lazy = FALSE))
+  expect_warning(out5 <- read_csv(I("x1,x2,x3\nA2,B2,C2\nA3#,B2,C2\nA4,A5,A6"), comment = "#", lazy = FALSE))
+  expect_warning(out6 <- read_csv(I("x1,x2,x3\nA2,B2,C2\nA3,#B2,C2\nA4,A5,A6"), comment = "#", lazy = FALSE))
+  expect_warning(out7 <- read_csv(I("x1,x2,x3\nA2,B2,C2\nA3,#B2,C2\n#comment\nA4,A5,A6"), comment = "#", lazy = FALSE))
 
   chk <- tibble::tibble(
     x1 = c("A2", "A3", "A4"),
@@ -233,15 +233,15 @@ test_that("comments are ignored regardless of where they appear", {
 
 test_that("escaped/quoted comments are ignored", {
   out1 <- read_delim(I('x\n\\#'), comment = "#", delim = ",",
-    escape_backslash = TRUE, escape_double = FALSE, progress = FALSE)
-  out2 <- read_csv(I('x\n"#"'), comment = "#", progress = FALSE)
+    escape_backslash = TRUE, escape_double = FALSE)
+  out2 <- read_csv(I('x\n"#"'), comment = "#")
 
   expect_equal(out1$x, "#")
   expect_equal(out2$x, "#")
 })
 
 test_that("leading comments are ignored", {
-  out <- read_csv(I("#a\n#b\nx\n1"), comment = "#", progress = FALSE)
+  out <- read_csv(I("#a\n#b\nx\n1"), comment = "#")
 
   expect_equal(ncol(out), 1)
   expect_equal(out$x, 1L)
@@ -249,7 +249,7 @@ test_that("leading comments are ignored", {
 
 test_that("skip respects comments", {
   read_x <- function(...) {
-    read_csv(I("#a\nb\nc"), col_names = FALSE, ..., progress = FALSE)[[1]]
+    read_csv(I("#a\nb\nc"), col_names = FALSE, ...)[[1]]
   }
 
   expect_equal(read_x(), c("#a", "b", "c"))
@@ -261,7 +261,7 @@ test_that("skip respects comments", {
 test_that("skip respects newlines", {
   skip_if_edition_second()
   read_x <- function(...) {
-    read_csv(I("1\n2\n3\n\na\nb\nc"), col_names = FALSE, ..., progress = FALSE)[[1]]
+    read_csv(I("1\n2\n3\n\na\nb\nc"), col_names = FALSE, ...)[[1]]
   }
 
   expect_equal(read_x(), c("1", "2", "3", "a", "b", "c"))
@@ -277,7 +277,7 @@ test_that("skip respects newlines", {
 })
 
 test_that("read_csv returns an empty data.frame on an empty file", {
-   expect_equal(read_csv(test_path("empty-file"), progress = FALSE)[], tibble::tibble())
+   expect_equal(read_csv(test_path("empty-file"))[], tibble::tibble())
 })
 
 test_that("read_delim errors on length 0 delimiter (557)", {