+++ tests/testthat/test-write-lines.R
@@ -4,7 +4,7 @@ test_that("write_lines uses UTF-8 encoding", {
   tmp <- tempfile()
   on.exit(unlink(tmp))
   write_lines(c("fran\u00e7ais", "\u00e9l\u00e8ve"), tmp)
-  x <- read_lines(tmp, locale = locale(encoding = "UTF-8"), progress = FALSE)
+  x <- read_lines(tmp, locale = locale(encoding = "UTF-8"))
   expect_equal(x, c("fran\u00e7ais", "\u00e9l\u00e8ve"))
 })